refractory code of LearnBox component
